% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acidente_cnae.R
\name{acidente_cnae}
\alias{acidente_cnae}
\title{Function acidente_cnae}
\usage{
acidente_cnae(cnae)
}
\arguments{
\item{cnae}{vector of Brazilian Code of Economic Activity corresponding to the ISIC}
}
\value{
A data.frame with 7 variables with the information available.
}
\description{
This function returns
}
\examples{
Extract information from chemical, pharmaceutical and rubber industries.
df<-acidente_cnae(cnae=c(20,21,22)) 
}
\keyword{accidents}
\keyword{labor}
