% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediador_inteiro_teor.R
\name{mediador_inteiro_teor}
\alias{mediador_inteiro_teor}
\title{Function  mediador_inteiro_teor}
\usage{
mediador_inteiro_teor(solicitacao, download = TRUE, vector = TRUE)
}
\arguments{
\item{solicitacao}{number that works as an id of the document.}

\item{download}{whether to download the document or not}

\item{vector}{wheter to create a vector of the documents as an R object.}
}
\value{
A file with the document and/or an R object with the text.
}
\description{
This function downloads the whole texts of the agreements according to "solitacao" number
extracted through mediador_meta function
}
\examples{
m<-mediador_inteiro_teor(solicitacao=c("MR087305/2016","MR035244/2009")) 
}
\keyword{agreements.}
\keyword{labor}
\keyword{mediador,}
