% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediador_meta.R
\name{mediador_meta}
\alias{mediador_meta}
\title{Function mediador_meta}
\usage{
mediador_meta(livre = "", uf = "", cnpj = "")
}
\arguments{
\item{livre}{words to searched.}

\item{uf}{federative unit (state)}

\item{cnpj}{company or organization tax number}
}
\value{
A data.frame with metadata about the agreement.
}
\description{
This function returns a data.frame with information about collective agreements between workers' unions
and companies or companies' unions
}
\examples{
mediador_meta(livre="industria qumicia",uf="SP") 
}
\keyword{agreements}
\keyword{labor}
\keyword{mediador,}
