% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rais.R
\name{rais}
\alias{rais}
\title{Function rais}
\usage{
rais(uf = NULL, CNAE = "", CBO = "", year = NULL)
}
\arguments{
\item{uf}{vector of Brazilian federation unites (states + federal district)}

\item{CNAE}{Brazilian Code of Economic Activity corresponding to ISIC}

\item{CBO}{Brazilian Standard Classification of Ocupations}

\item{year}{year of data to download.}
}
\value{
A data.frame with
}
\description{
This function returns microdata of Brazilian workers
}
\examples{
rais(uf=c("AC","SC"),CNAE="22",CBO="21",year=2014)
}
\keyword{employment,}
\keyword{jobs,}
\keyword{ocupations,}
\keyword{rais,}
\keyword{workers}
